requirements.txt
Flask==3.1.0
mediapipe==0.10.14
opencv-python==4.10.0.84
paho-mqtt==2.1.0
requests==2.32.3
